Minibench refactor
Allow running generic model benchmark before LLM
