+++ extension/benchmark/android/benchmark/build.gradle.kts
@@ -7,4 +7,6 @@
  */
 
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
-plugins { id("com.android.application") version "8.1.0" apply false }
+plugins { id("com.android.application") version "8.1.0" apply false
+    id("org.jetbrains.kotlin.android") version "2.1.10" apply false
+}
